+++ src/main/java/com/flow/flow/controller/DetailController.java
@@ -1,0 +1,15 @@
+package com.flow.flow.controller;
+
+import com.flow.flow.service.DetailService;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(value = "")
+public class DetailController {
+    private DetailService detailService;
+
+    public DetailController(DetailService detailService) {
+        this.detailService = detailService;
+    }
+}
