+++ src/main/java/com/flow/flow/controller/KeywordController.java
@@ -1,0 +1,15 @@
+package com.flow.flow.controller;
+
+import com.flow.flow.service.KeywordService;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(value = "")
+public class KeywordController {
+    private KeywordService keywordService;
+
+    public KeywordController(KeywordService keywordService) {
+        this.keywordService = keywordService;
+    }
+}
