+++ src/main/java/com/flow/flow/controller/UserController.java
@@ -1,0 +1,15 @@
+package com.flow.flow.controller;
+
+import com.flow.flow.service.UserService;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping(value = "/users")
+public class UserController {
+    private UserService userService;
+
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+}
