--- src/main/java/com/flow/flow/detail/dto/DetailDto.java
@@ -1,4 +1,0 @@
-package com.flow.flow.detail.dto;
-
-public class DetailDto {
-}
