--- src/main/java/com/flow/flow/detail/repository/DetailRepository.java
@@ -1,9 +1,0 @@
-package com.flow.flow.detail.repository;
-
-import com.flow.flow.detail.entity.Detail;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface DetailRepository extends JpaRepository<Detail, Long> {
-}
