+++ src/main/java/com/flow/flow/domain/Form.java
@@ -1,0 +1,32 @@
+package com.flow.flow.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Form {
+
+    @Id @GeneratedValue
+    private Long formId;
+
+    private String bankName;
+
+    private boolean isExp;
+
+    private String form;
+
+    @Builder
+    public Form(String bankName, boolean isExp, String form
+    ) {
+        this.bankName = bankName;
+        this.isExp = isExp;
+        this.form = form;
+    }
+}
