+++ src/main/java/com/flow/flow/dto/DetailDto.java
@@ -1,0 +1,4 @@
+package com.flow.flow.dto;
+
+public class DetailDto {
+}
