+++ src/main/java/com/flow/flow/dto/KeywordDto.java
@@ -1,0 +1,4 @@
+package com.flow.flow.dto;
+
+public class KeywordDto {
+}
