+++ src/main/java/com/flow/flow/dto/UserDto.java
@@ -1,0 +1,4 @@
+package com.flow.flow.dto;
+
+public class UserDto {
+}
