--- src/main/java/com/flow/flow/form/dto/FormDto.java
@@ -1,4 +1,0 @@
-package com.flow.flow.form.dto;
-
-public class FormDto {
-}
