--- src/main/java/com/flow/flow/form/repository/FormRepository.java
@@ -1,9 +1,0 @@
-package com.flow.flow.form.repository;
-
-import com.flow.flow.form.entity.Form;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface FormRepository extends JpaRepository<Form, Long> {
-}
