--- src/main/java/com/flow/flow/form/service/FormService.java
@@ -1,13 +1,0 @@
-package com.flow.flow.form.service;
-
-import com.flow.flow.form.repository.FormRepository;
-import org.springframework.stereotype.Service;
-
-@Service
-public class FormService {
-    private FormRepository formRepository;
-
-    public FormService(FormRepository formRepository) {
-        this.formRepository = formRepository;
-    }
-}
