--- src/main/java/com/flow/flow/keyword/controller/KeywordController.java
@@ -1,15 +1,0 @@
-package com.flow.flow.keyword.controller;
-
-import com.flow.flow.keyword.service.KeywordService;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(value = "")
-public class KeywordController {
-    private KeywordService keywordService;
-
-    public KeywordController(KeywordService keywordService) {
-        this.keywordService = keywordService;
-    }
-}
