--- src/main/java/com/flow/flow/keyword/dto/KeywordDto.java
@@ -1,4 +1,0 @@
-package com.flow.flow.keyword.dto;
-
-public class KeywordDto {
-}
