--- src/main/java/com/flow/flow/keyword/entity/Keyword.java
@@ -1,24 +1,0 @@
-package com.flow.flow.keyword.entity;
-
-import com.flow.flow.user.entity.User;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class Keyword {
-
-    @Id @GeneratedValue
-    private Long keywordId;
-
-    private String keyword;
-
-    // category id
-
-    @OneToOne
-    @JoinColumn(name = "userId")
-    private User user;
-}
