--- src/main/java/com/flow/flow/keyword/repository/KeywordRepository.java
@@ -1,9 +1,0 @@
-package com.flow.flow.keyword.repository;
-
-import com.flow.flow.keyword.entity.Keyword;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface KeywordRepository extends JpaRepository<Keyword, Long> {
-}
