+++ src/main/java/com/flow/flow/repository/DetailRepository.java
@@ -1,0 +1,9 @@
+package com.flow.flow.repository;
+
+import com.flow.flow.domain.Detail;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface DetailRepository extends JpaRepository<Detail, Long> {
+}
