+++ src/main/java/com/flow/flow/repository/FormRepository.java
@@ -1,0 +1,9 @@
+package com.flow.flow.repository;
+
+import com.flow.flow.domain.Form;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface FormRepository extends JpaRepository<Form, Long> {
+}
