+++ src/main/java/com/flow/flow/repository/KeywordRepository.java
@@ -1,0 +1,9 @@
+package com.flow.flow.repository;
+
+import com.flow.flow.domain.Keyword;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface KeywordRepository extends JpaRepository<Keyword, Long> {
+}
