+++ src/main/java/com/flow/flow/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.flow.flow.repository;
+
+import com.flow.flow.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+}
