+++ src/main/java/com/flow/flow/service/UserService.java
@@ -1,0 +1,13 @@
+package com.flow.flow.service;
+
+import com.flow.flow.repository.UserRepository;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+    private UserRepository userRepository;
+
+    public UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+}
