--- src/main/java/com/flow/flow/user/controller/UserController.java
@@ -1,15 +1,0 @@
-package com.flow.flow.user.controller;
-
-import com.flow.flow.user.service.UserService;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping(value = "/users")
-public class UserController {
-    private UserService userService;
-
-    public UserController(UserService userService) {
-        this.userService = userService;
-    }
-}
