--- src/main/java/com/flow/flow/user/dto/UserDto.java
@@ -1,4 +1,0 @@
-package com.flow.flow.user.dto;
-
-public class UserDto {
-}
