--- src/main/java/com/flow/flow/user/entity/User.java
@@ -1,44 +1,0 @@
-package com.flow.flow.user.entity;
-
-import com.flow.flow.detail.entity.Detail;
-import com.flow.flow.keyword.entity.Keyword;
-import lombok.*;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class User {
-
-    @Id @GeneratedValue
-    private Long userId;
-
-    private String email;
-
-    private String userName;
-
-    private String socialId;
-
-    @Enumerated(EnumType.STRING)
-    private SocialType socialType;
-
-    @OneToOne(mappedBy = "user")
-    private Budget budget;
-
-    @OneToMany(mappedBy = "user")
-    private List<Detail> details = new ArrayList<>();
-
-    @OneToOne(mappedBy = "user")
-    private Keyword keyword;
-
-    @Builder
-    public User(String email, String userName, String socialId, SocialType socialType) {
-        this.email = email;
-        this.userName = userName;
-        this.socialId = socialId;
-        this.socialType = socialType;
-    }
-}
