--- src/main/java/com/flow/flow/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.flow.flow.user.repository;
-
-import com.flow.flow.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
