--- src/main/java/com/flow/flow/user/service/UserService.java
@@ -1,13 +1,0 @@
-package com.flow.flow.user.service;
-
-import com.flow.flow.user.repository.UserRepository;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-    private UserRepository userRepository;
-
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-}
